requirements.txt
flask 
flask-socketio 
python-chess
openai